refactor: move handlers to server dir
