+++ cmd/service/main.go
@@ -12,10 +12,10 @@ import (
 	"github.com/nix-united/golang-gin-boilerplate/docs"
 	"github.com/nix-united/golang-gin-boilerplate/internal/config"
 	"github.com/nix-united/golang-gin-boilerplate/internal/db"
-	"github.com/nix-united/golang-gin-boilerplate/internal/handler"
 	"github.com/nix-united/golang-gin-boilerplate/internal/provider"
 	"github.com/nix-united/golang-gin-boilerplate/internal/repository"
 	"github.com/nix-united/golang-gin-boilerplate/internal/server"
+	"github.com/nix-united/golang-gin-boilerplate/internal/server/handler"
 	"github.com/nix-united/golang-gin-boilerplate/internal/service/post"
 	"github.com/nix-united/golang-gin-boilerplate/internal/service/user"
 	"github.com/nix-united/golang-gin-boilerplate/internal/utils"
