+++ internal/server/handler/auth_handler_mock_test.go
